backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
anthropic
elevenlabs
pydub
librosa
soundfile
supabase
phonemizer
